feat: dist-tags configuration
